updated functions with async/await

+++ server/server.js
@@ -29,76 +29,68 @@ app.get('/', (req, res) => {
 })
 
 // Post Todo
-app.post('/todos', authenticate, (req, res) => {
-  const todo = new Todo({
-    text: req.body.text,
-    _creator: req.user._id
-  })
+app.post('/todos', authenticate, async (req, res) => {
+  const todo = new Todo({ text: req.body.text, _creator: req.user._id })
 
-  todo.save().then((doc) => {
+  try {
+    const doc = await todo.save()
     res.send(doc)
-  }, (err) => {
-    res.status(400).send(err)
-  })
+  } catch (error) {
+    res.status(400).send(error)
+  }
 })
 
 // Get all Todos
-app.get('/todos', authenticate, (req, res) => {
-  Todo.find({
-    _creator: req.user._id
-  }).then((todos) => {
+app.get('/todos', authenticate, async (req, res) => {
+  try {
+    const todos = await Todo.find({ _creator: req.user._id })
     res.send({todos})
-  }, (err) => {
-    res.status(400).send(err)
-  })
+  } catch (error) {
+    res.status(400).send(error)
+  }
 })
 
 // Get Todo by Id
-app.get('/todos/:id', authenticate, (req, res) => {
+app.get('/todos/:id', authenticate, async (req, res) => {
   const id = req.params.id
 
   if (!ObjectId.isValid(id)) {
     return res.status(404).send('Invalid ObjectId')
   }
 
-  Todo.findOne({
-    _id: id,
-    _creator: req.user._id
-  }).then((todo) => {
+  try {
+    const todo = await Todo.findOne({ _id: id, _creator: req.user._id })
     if (!todo) {
       return res.status(404).send('Todo not found')
     }
 
-    res.send({todo})
-  }).catch((err) => {
+    res.send({ todo })
+  } catch (error) {
     res.status(400).send('Database Error')
-  })
+  }
 })
 
 // Delete Todo by Id
-app.delete('/todos/:id', authenticate, (req, res) => {
-  const id = req.params.id
+app.delete('/todos/:id', authenticate, async (req, res) => {
+    const id = req.params.id
+    if (!ObjectId.isValid(id)) {
+      return res.status(404).send('Invalid ObjectId')
+    }
 
-  if (!ObjectId.isValid(id)) {
-    return res.status(404).send('Invalid ObjectId')
-  }
-  
-  Todo.findOneAndRemove({
-    _id: id,
-    _creator: req.user._id
-  }).then((todo) => {
+  try {
+    const todo = await Todo.findOneAndRemove({ _id: id, _creator: req.user._id })
     if (!todo) {
       return res.status(404).send('Todo not found')
     }
 
-    res.send({todo})
-  }).catch((err) => {
+    res.send({ todo })
+  } catch (error) {
     res.status(400).send('Database Error')
-  })
+  }
 })
 
 // Patch Todo
-app.patch('/todos/:id', authenticate, (req, res) => {
+app.patch('/todos/:id', authenticate, async (req, res) => {
   const id = req.params.id
   const body = _.pick(req.body, ['text', 'completed'])
 
@@ -113,32 +105,32 @@ app.patch('/todos/:id', authenticate, (req, res) => {
     body.completedAt = null
   }
 
-  Todo.findOneAndUpdate({
-    _id: id,
-    _creator: req.user._id
-  }, {$set: body}, {new: true}).then((todo) => {
+  try {
+    const todo = await Todo.findOneAndUpdate({
+      _id: id,
+      _creator: req.user._id
+    }, {$set: body}, {new: true})
     if (!todo) {
       return res.status(404).send('Todo not found')
     }
 
     res.send({todo})
-  }).catch((err) => {
+  } catch (error) {
     res.status(400).send('Database Error')
-  })
+  }
 })
 
 // POST /users
-app.post('/users', (req, res) => {
-  const body = _.pick(req.body, ['email', 'password'])
-  const user = new User(body)
-
-  user.save().then(() => {
-    return user.generateAuthToken()
-  }).then((token) => {
+app.post('/users', async (req, res) => {
+  try {
+    const body = _.pick(req.body, ['email', 'password'])
+    const user = new User(body)
+    await user.save()
+    const token = await user.generateAuthToken()
     res.cookie('token', token).send(user)
-  }).catch((err) => {
-    res.status(400).send(err)
-  })
+  } catch (error) {
+    res.status(400).send(error)
+  }
 })
 
 // Private route
@@ -147,25 +139,25 @@ app.get('/users/me', authenticate, (req, res) => {
 })
 
 // POST /users/login
-app.post('/users/login', (req, res) => {
-  const body = _.pick(req.body, ['email', 'password'])
-
-  User.findByCredentials(body.email, body.password).then((user) => {
-    return user.generateAuthToken().then((token) => {
-      res.cookie('token', token).send(user)
-    })
-  }).catch((err) => {
+app.post('/users/login', async (req, res) => {
+  try {
+    const body = _.pick(req.body, ['email', 'password'])
+    const user = await User.findByCredentials(body.email, body.password)
+    const token = await user.generateAuthToken()
+    res.cookie('token', token).send(user)
+  } catch (error) {
     res.status(400).send('Invalid Login Credentials')
-  })
+  }
 })
 
 // Logout
-app.delete('/users/me/token', authenticate , (req, res) => {
-  req.user.removeToken(req.token).then(() => {
+app.delete('/users/me/token', authenticate, async (req, res) => {
+  try {
+    await req.user.removeToken(req.token)
     res.status(200).send()
-  }, () => {
+  } catch(err) {
     res.status(400).send()
-  })
+  }
 })
 
 // Start Server
